Move quota validator into dynamic
